chore(file name): rename file to follow Flutter code style
