+++ lib/services/characters/models/sex_enum.dart
