+++ lib/ui/home/home_page_bloc.dart
@@ -1,7 +1,7 @@
-import 'package:Sample_Characters/services/characters/repositories/CharacterRepository.dart';
+import 'package:Sample_Characters/services/characters/repositories/character_repository.dart';
 import 'package:get_it/get_it.dart';
-import 'package:Sample_Characters/services/characters/models/Character.dart';
-import 'package:Sample_Characters/ui/common/base_classes/blocs/BaseBloc.dart';
+import 'package:Sample_Characters/services/characters/models/character.dart';
+import 'package:Sample_Characters/ui/common/base_classes/blocs/base_bloc.dart';
 
 class HomePageBloc implements BaseBloc {
   HomePageBloc();
